Bug 494639 - NJ: fix numerous LIR memory management problems, r=graydon

+++ js/src/nanojit/Native.h
@@ -60,7 +60,7 @@
 #endif
 
 namespace nanojit {
-	const uint32_t NJ_PAGE_SIZE = 1 << NJ_LOG2_PAGE_SIZE;
+    const size_t NJ_PAGE_SIZE = 1 << NJ_LOG2_PAGE_SIZE;
 	
     class Fragment;
     struct SideExit;
